IN3110_assignments/IN3110_assignment3/pyproject.toml
[build-system]
requires = [
    "setuptools",
    #4110 only:
    #"cython",
    #"numpy==1.21.*",
]
build-backend = "setuptools.build_meta"

[project]
version = "0.1.0"
requires-python = ">=3.7"
license = {text = "MIT License"}
name = "instapy"
description = "Turn images to dramatic grayscale or nostalgic sepia image"
readme = "README.md"
dependencies = [
    "numpy",
    "numba",
    "pillow",
    "line-profiler"
 ]


[project.scripts]
#instapy = "instapy.__main__:main"
